Move safe_hdf5 into new h5_utils module
Having a h5_utils header+implementation module is in general probably a
good idea, as it will help moving some of the more general utilities out
of hdfitems.h, which currently has defined loads of code inline.
